+++ app/src/main/res/layout/activity_login.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="viewmodel"
-            type="br.com.aisdigital.androidchallenge.LoginViewModel" />
+            type="br.com.aisdigital.androidchallenge.ui.activities.login.LoginViewModel" />
     </data>
 
     <ScrollView
@@ -23,7 +23,7 @@
             android:paddingTop="@dimen/padding_default"
             android:paddingRight="@dimen/padding_default"
             android:paddingBottom="@dimen/padding_default"
-            tools:context=".ui.LoginActivity">
+            tools:context=".ui.activities.login.LoginActivity">
 
             <LinearLayout
                 android:id="@+id/formLogin"
